Debugged key exp, enc, and dec pipes, BH
Debugged the key expansion system.  Rcon was off by a clock and sbox
data was off by a clock.  Also like Florian originally said, there are
11 keys in the expansion, so I added that.
Debugged the ENC, DEC, and Stream In blocks.  Minor changes to ENC.  DEC
Inv Mix Col needed some major rework.  It is not a less-compressed
version that was originally thought.

+++ runsim.pl
@@ -7,13 +7,19 @@ use File::Basename;
 
 
 my $stim;
+my $tb;
+my $nosim;
 my $gui;
 my $clean;
+my $waves;
 my $help;
 
 GetOptions ('help'     => \$help,
+            'tb=s'     => \$tb,
+            'nosim'    => \$nosim,
             'stim=s'   => \$stim,
             'clean'    => \$clean,
+            'waves'    => \$waves,
             'gui'      => \$gui);
 
 
@@ -23,8 +29,11 @@ USAGE
       runsim.pl -stim <FILE>
 
 OPTIONS
+      -tb   <FILE>         Testbench to run.
       -stim <FILE>         Stimulus file to use.
+      -nosim               Don't simulate, just print files.
       -gui                 Run sim with gui mode.
+      -waves               Dumps waves with the sim.
       -clean               Removes all tmp files.
       -help                Prints this usage.
 
@@ -49,14 +58,14 @@ if (defined $clean) {
 
 
 # Copy the stim file to the tmp dir
+die "ERROR: No tb ($tb) exists!"     if (!-e $tb);
 die "ERROR: No stim ($stim) exists!" if (!-e $stim);
 system("cp $stim stim.v");
 
 
 # Get a list of files to compile
-my @files = ();
-my @dirs  = ('../rtl',
-             '../tb');
+my @files = ($tb);
+my @dirs  = ('../rtl');
 
 foreach my $dir (@dirs) {
   opendir DIR, $dir;
@@ -72,16 +81,23 @@ foreach my $dir (@dirs) {
 
 
 # Run the NC-Verilog command
-my $cmd = 'ncverilog +sv +nc64bit ';
+my $cmd = 'ncverilog +sv +nc64bit +define+USING_RUNSIM ';
 foreach my $file (@files) {
   $cmd .= " $file";
 }
-$cmd .= ' -clean' if (defined $clean);
-$cmd .= ' +define+DUMP_WAVES +access+r +gui &' if (defined $gui);
+$waves = 1 if (defined $gui);
+$cmd .= ' -clean'                       if (defined $clean);
+$cmd .= ' +define+DUMP_WAVES +access+r' if (defined $waves);
+$cmd .= ' +gui &'                       if (defined $gui);
 print "\n\nRUNNING COMMAND: '$cmd'\n";
 print "\n-------------------- NC Verilog --------------------\n";
-system($cmd);
 
+if (defined $nosim) {
+  print "NO SIMULATION due to -nosim switch!\n";
+}
+else {
+  system($cmd);
+}
 
 print "\n---------------------------------------------------\n";
